server/requirements.txt
requests==2.20.1
Flask==1.1.2
Werkzeug==1.0.1
beautifulsoup4==4.9.3
cloudinary==1.24.0
flask_cors==3.0.10
flask_restful==0.3.8
python-dotenv==0.15.0
